Modified for Assignment

+++ methods/StyleAdv_RN_GNN.py
@@ -27,7 +27,12 @@ class StyleAdvGNN(MetaTemplate):
 
     # fix label for training the metric function   1*nw(1 + ns)*nw
     support_label = torch.from_numpy(np.repeat(range(self.n_way), self.n_support)).unsqueeze(1)
-    support_label = torch.zeros(self.n_way*self.n_support, self.n_way).scatter(1, support_label, 1).view(self.n_way, self.n_support, self.n_way)
+    # support_label = torch.zeros(self.n_way*self.n_support, self.n_way).scatter(1, support_label, 1).view(self.n_way, self.n_support, self.n_way)
+    
+    support_label = torch.zeros(self.n_way*self.n_support, self.n_way).scatter(
+    1, support_label.long(), 1  
+    ).view(self.n_way, self.n_support, self.n_way)
+    
     support_label = torch.cat([support_label, torch.zeros(self.n_way, 1, n_way)], dim=1)
     self.support_label = support_label.view(1, -1, self.n_way)
